move tokenize utils out into a sep file
